+++ src/group_by.cpp
@@ -11,26 +11,9 @@ using namespace Rcpp;
 using namespace dplyr;
 
 // [[Rcpp::export]]
-DataFrame grouped_df_impl(DataFrame data, SymbolVector symbols, bool build_index = true) {
-  assert_all_white_list(data);
-  DataFrame copy(shallow_copy(data));
-  set_vars(copy, symbols);
-  set_class(copy, classes_grouped<GroupedDataFrame>());
-  if (!symbols.size())
-    stop("no variables to group by");
-  if (build_index) {
-    build_index_cpp(copy);
-  }
-  else {
-    strip_index(copy);
-  }
-  return copy;
-}
-
-// [[Rcpp::export]]
 DataFrame as_regular_df(DataFrame df) {
   DataFrame copy(shallow_copy(df));
-  SET_ATTRIB(copy, strip_group_attributes(df));
+  copy.attr("groups") = R_NilValue ;
   SET_OBJECT(copy, OBJECT(df));
   set_class(copy, CharacterVector::create("data.frame"));
   return copy;
@@ -39,7 +22,8 @@ DataFrame as_regular_df(DataFrame df) {
 // [[Rcpp::export]]
 DataFrame ungroup_grouped_df(DataFrame df) {
   DataFrame copy(shallow_copy(df));
-  SET_ATTRIB(copy, strip_group_attributes(df));
+  copy.attr("groups") = R_NilValue ;
+  set_class(copy, classes_not_grouped());
   return copy;
 }