+++ tests/testthat/test-binds.R
@@ -490,7 +490,7 @@ test_that("bind_rows infers classes from first result (#1692)", {
   expect_equal(class(bind_rows(d2, d1)), c("tbl_df", "tbl", "data.frame"))
   res3 <- bind_rows(d3, d1)
   expect_equal(class(res3), c("grouped_df", "tbl_df", "tbl", "data.frame"))
-  expect_equal(attr(res3, "group_sizes"), c(10, 10))
+  expect_equal(map_int(group_rows(res3), length), c(10, 10))
   expect_equal(class(bind_rows(d4, d1)), c("rowwise_df", "tbl_df", "tbl", "data.frame"))
 
   expect_equal(class(bind_rows(d5, d1)), c("tbl_df", "tbl", "data.frame"))
@@ -507,7 +507,7 @@ test_that("bind_cols infers classes from first result (#1692)", {
   expect_equal(class(bind_cols(d2, d1)), c("tbl_df", "tbl", "data.frame"))
   res3 <- bind_cols(d3, d1)
   expect_equal(class(res3), c("grouped_df", "tbl_df", "tbl", "data.frame"))
-  expect_equal(attr(res3, "group_sizes"), c(5, 5))
+  expect_equal(map_int(group_rows(res3), length), c(5, 5))
   expect_equal(class(bind_cols(d4, d1)), c("rowwise_df", "tbl_df", "tbl", "data.frame"))
   expect_equal(class(bind_cols(d5, d1)), c("tbl_df", "tbl", "data.frame"))
 })